lógica de negocio completa utilizando implementando todo lo anterior y usando MediaTR, asi como utilizando patron repositorio generico utilizado como bibloteca de clases e implementado en la API unicamente

+++ Core/Interfaces/Repository/IRepository.cs
@@ -5,14 +5,14 @@ namespace Core.Interfaces.Repository
     public interface IRepository<T> where T : class
     {
         //Consultas
-        Task<T?> GetByIdAsync(long id, CancellationToken cancellationToken = default);
+        Task<T?> GetByIdAsync(int id, CancellationToken cancellationToken = default);
         Task<IReadOnlyList<T>> GetAllAsync(CancellationToken cancellationToken);
         Task<IReadOnlyList<T>> GetByFilterAsync(Expression<Func<T, bool>> filter, Expression<Func<T, object>>? orderBy = null,
         bool ascending = true, CancellationToken cancellationToken = default);
         Task<bool> ExistsAsync(Expression<Func<T, bool>> predicate, CancellationToken cancellationToken = default);
 
         //Comandos
-        Task<bool> DeleteAsync(long id, CancellationToken cancellationToken);
+        Task<bool> DeleteAsync(int id, CancellationToken cancellationToken);
         Task AddAsync(T entity, CancellationToken cancellationToken);
         Task<bool> UpdateAsync(T entity, CancellationToken cancellationToken);
         Task<string> GenerateNextCodigoAsync(CancellationToken cancellationToken);
